Implementando um filtro para a busca de PC

+++ src/main/java/br/com/inventarioApi/service/ComputadorService.java
@@ -2,7 +2,9 @@ package br.com.inventarioApi.service;
 
 import java.util.List;
 
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import br.com.inventarioApi.model.Computador;
@@ -13,21 +15,55 @@ public class ComputadorService {
 	
 	@Autowired
 	private ComputadorRepository computadorRepository;
-	
-	
+		
 	//salvando e alterando
-	public void salvar(Computador computador) {
-		computadorRepository.save(computador);
+	public Computador salvar(Computador computador) {
+		return computadorRepository.save(computador);
 	}
 	
 	//listando
-	public List<Computador> listar(){
+	public List<Computador> listarComputadores(){
 		return computadorRepository.findAll();
 	}
 	
 	//deletando
 	public void deletar(Integer codigo) {
-		computadorRepository.deleteById(codigo);
+		
+		Computador pcPraDeletar = computadorRepository.findOne(codigo);
+		
+		if(pcPraDeletar == null) {
+			throw new EmptyResultDataAccessException(1);
+		}else {
+			computadorRepository.deleteById(codigo);
+		}
 	}
+	
+	//atualizando
+	public Computador atualizar(Integer codigo, Computador computador) {
+		Computador pcPraAtualizar = buscaComputadorPeloCodigo(codigo);
+		BeanUtils.copyProperties(computador, pcPraAtualizar,"id");
+		return computadorRepository.save(pcPraAtualizar);
+	}
+	
+	
+	public Computador buscaComputadorPeloCodigo(Integer codigo) {
+		Computador computadorEncontrado = computadorRepository.findOne(codigo);
+		if(computadorEncontrado == null) {
+			throw new EmptyResultDataAccessException(1);
+		}
+		return computadorEncontrado;
+	}
+	
+	
 
 }
+
+
+
+
+
+
+
+
+
+
